interfaces: Can't believe it's working now
Added better handling of the scene (with diferenciated calculation of
phisics and collissions due to them).

This allows us to implement Velocity Verlet method to model the system,
as seen in:
http://gamedev.stackexchange.com/questions/15708/how-can-i-implement-gravity

This commit also adds a simple configuration window that should be
easily extensible.

I maybe take the time to implement it with polygons, though torque could
potentially be difficult to implement.